+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/types/options/data/DraggableOptionEntry.java
@@ -1,0 +1,7 @@
+package com.teamresourceful.resourcefulconfig.api.types.options.data;
+
+public record DraggableOptionEntry<T>(
+        T value,
+        boolean duplicatable
+) {
+}